king: fromNoun should always force result, so add NFData constraint.
fromNoun now requires that the type it parses to is an NFData so that
you can force at the end of it. Usually, data to parse has been read
as a bytestring off the wire, has been cued (lazily) and is then passed
to noun parsing, which was also lazy and would often keep the input
bytestrings for reasons I don't really understand.

So this the resulting parsed data structure at the end of noun parsing.
This sprinkles NFData instances throughout the codebase so everything
that we convert to a type which we can `force`.

For ease of use, this modifies the template haskell so that matching
`Generic` and NFData instances are added to any type where you call
`deriveNoun` on.